+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 1.15.5 - 2015-01-22
+ * rubocop changes
+
+## 1.15.4 - 2015-01-22
+ * updated the url for reply messages
+
+## 1.15.3 - 2015-01-22
+ * updated redirect when a participant resets password
+
 ## 1.15.2 - 2015-01-22
  * fixed sorting issues in patient dashboard